refactor wijzig periode dialoog
